requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.54.1
h5py==3.12.1
idna==3.10
importlib==1.0.4
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy==1.14.1
six==1.16.0
tabulate==0.9.0
tqdm==4.66.5
tzdata==2024.2
urllib3==2.2.3
wheel==0.44.0
